Fix comments, use Wrapper class, instead of primitive type. Change to double where required

+++ src/main/java/project/financemanagement/demo/model/Budget.java
@@ -17,11 +17,11 @@ import java.time.LocalDateTime;
 public class Budget {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private long budgetId; /*a unique identifier for the budget*/
+    private Long budgetId; /*a unique identifier for the budget*/
     private String budgetCategory; /*the category the budget applies to (e.g. rent, groceries, entertainment)*/
-    private long budgetAmount; /*the budgeted amount for the category*/
+    private double budgetAmount; /*the budgeted amount for the category*/
     private LocalDateTime startDate;  /*the start date of the budget period*/
     private LocalDateTime endDate;  /*the end date of the budget period*/
-    private long goal; /*a financial goal associated with the budget*/
-    private long progress; /*the current progress towards the goal*/
+    private double goal; /*a financial goal associated with the budget*/
+    private double progress; /*the current progress towards the goal*/
 }
