+++ src/main/java/project/financemanagement/demo/model/FinancialAccount.java
@@ -14,10 +14,10 @@ import java.util.List;
 public class FinancialAccount {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private long accountId; /*a unique identifier for the account*/
+    private Long accountId; /*a unique identifier for the account*/
     private String accountName; /*a name for the account (e.g. savings account, 401k)*/
     private String accountType; /*the type of account (e.g. savings, checking, investment)*/
-    private long accountBalance; /*the current balance of the account*/
+    private double accountBalance; /*the current balance of the account*/
 
     @OneToMany(mappedBy = "financialAccount", cascade = CascadeType.ALL)
     @ToString.Exclude
